+++ src/app/xm-dashboard/page/index.ts
@@ -1,1 +1,0 @@
-export { PageService } from './page.service';